+++ CHANGELOG.md
@@ -56,6 +56,9 @@ Not quite compatible with Octave yet. See #43 [octave support](https://github.co
 * Fix a tool crash in MH Metrics: when producing the HTML report the
   tool would crash if there were parse errors in at least one file.
 
+* MH Style support a new rule "naming_parameters" that enforces a
+  naming scheme for function and method inputs and outputs.
+
 ### 0.9.22
 
 * Fix bug in MH Trace: libraries unit tests were considered part of